debug

+++ db_init.sql
@@ -3,7 +3,7 @@ CREATE TABLE raw_data (
     student VARCHAR(32),
     student_id INT,
     subject VARCHAR(32),
-    score DOUBLE,
+    score REAL,
     file_url VARCHAR(512),
     feedback VARCHAR(4096),
     submit_time INT,
